Refactor: 타입 관련 리팩토링
