+++ src/components/mainPage/videoGrid/VideoGrid.tsx
@@ -1,18 +1,18 @@
 import React from "react";
 import VideoCard from "../videoCard/VideoCard";
-import { VideoGridProps } from "../videoGrid/types";
 import { GridContainer } from "../videoGrid/styles";
+import { Video } from "@@types/video.type";
 
-const VideoGrid: React.FC<VideoGridProps> = ({ videos, onVideoClick, size = "medium" }) => {
+export interface VideoGridProps {
+    videos: Video[];
+    size?: "small" | "medium" | "large";
+}
+
+const VideoGrid: React.FC<VideoGridProps> = ({ videos, size = "medium" }) => {
     return (
         <GridContainer>
             {videos.map((video) => (
-                <VideoCard
-                    key={video.id}
-                    video={video}
-                    size={size}
-                    onClick={() => onVideoClick?.(video)}
-                />
+                <VideoCard key={video.videopostId} video={video} size={size} />
             ))}
         </GridContainer>
     );
