--- src/components/mainPage/videoGrid/types.ts
@@ -1,7 +1,0 @@
-import { Video } from '../../../types/video.type';
-
-export interface VideoGridProps {
-    videos: Video[];
-    onVideoClick?: (video: Video) => void;
-    size?: 'small' | 'medium' | 'large';
-}
